Don't leak descriptors/handles for synchronous processes.
Instances of the various Process* classes have always survived for the
duration of the JVM, so they've never had any need to close the various
descriptors/handles they use for their processing loops.

Now, with synchronous processes, Process* instances can be used in a
one-shot manner, where they run a processing loop for a single process
and are then discarded. For that use case, after the processing loop is
completed, descriptors/handles need to be closed to prevent leaks.

- Updated ProcessCompletions, ProcessEpoll and ProcessKqueue to close
  their respective descriptors/handles after synchronous processes end