Null injection hardening
In NuProcessBuilder throw an IllegalArgumentException if commands that
include a null character are passed. This hardens against command line
injection type attacks in applications that use NuProcess. Throwing a
IllegalArgumentException is what Java's ProcessBuilder does when nulls
are passed so this solution was chosen to maintain some consistency with
that.