Remove synchronization when starting Linux processes.
Since NuProcess on Linux made the switch to using the JVM's own
process start logic, it's likely this synchronization is just a
vestigial safeguard.

Starting processes with run(), on Linux, has always allowed for
concurrent launches, and some clients use that heavily. That
reinforces the sense that the start() synchronization isn't
actually necessary.

This removes the synchronized block and adds a heavily threaded
test when processes are started concurrently.

fixes #159